requirements.txt
json
streamlit
request